+++ app/src/mobile/java/com/ssmhdssmhd/mxboxs/ui/fragment/SettingFragment.java
@@ -134,6 +134,7 @@ public class SettingFragment extends BaseFragment implements ConfigListener, Sit
         mBinding.danmaku.setOnClickListener(this::onDanmaku);
         mBinding.restore.setOnClickListener(this::onRestore);
         mBinding.version.setOnClickListener(this::onVersion);
+        mBinding.version.setOnLongClickListener(this::onVersionMirror);
         mBinding.parseServer.setOnClickListener(this::onParseServer);
         mBinding.parseServer.setOnLongClickListener(this::onParseServerReset);
         mBinding.vod.setOnLongClickListener(this::onVodEdit);
@@ -273,6 +274,11 @@ public class SettingFragment extends BaseFragment implements ConfigListener, Sit
 
     private void onVersion(View view) {
         Updater.create().force().start(requireActivity());
+    }
+
+    private boolean onVersionMirror(View view) {
+        Updater.create().showMirrorDialog(requireActivity());
+        return true;
     }
 
     private void setWallDefault(View view) {